mend

+++ engine/modules/graphics/internal/texturearray/array.go
@@ -4,10 +4,10 @@ import "github.com/go-gl/gl/v4.5-core/gl"
 
 type textureArray struct {
 	texture     uint32
-	imagesCount int
+	imagesCount int16
 }
 
-func (s *textureArray) Texture() uint32  { return s.texture }
-func (s *textureArray) ImagesCount() int { return s.imagesCount }
-func (s *textureArray) Release()         { gl.DeleteTextures(1, &s.texture) }
-func (s *textureArray) Bind()            { gl.BindTexture(gl.TEXTURE_2D_ARRAY, s.texture) }
+func (s *textureArray) Texture() uint32    { return s.texture }
+func (s *textureArray) ImagesCount() int16 { return s.imagesCount }
+func (s *textureArray) Release()           { gl.DeleteTextures(1, &s.texture) }
+func (s *textureArray) Bind()              { gl.BindTexture(gl.TEXTURE_2D_ARRAY, s.texture) }
